Factorize render function.
It now has parameter r a request pointer and from it, it retrieves the context, the user data and the main funcs needed in the funcsMap such a LoggedIn. Moved LoggedIn function to helpers due to cyclic graph between controllers and helpers. Clean up code that creates user data and loggedIn functions. Also handle case when func is null.

Fixes #51

Fixes #53